Fixed up the db meter graph. Source-level compiler warnings now use the following format: #pragma message cwarn("message"). Added CPL_RUNTIME_EXCEPTION() which throws and debug breaks a message that holds file, function and line number. Some api functions, and a breaking change in the cpl::ProgramInfo struct - user can now specify file directory. The complex resonator now has correct tuning and coefficients or windows (it still seems slightly off, though). Frequencygraph now supports linear scaling, presetwidgets are added together with md5 checked serialization. Openglimage touched up a lot. Added lock free spsc queues.

+++ CThread.h
@@ -167,12 +167,12 @@
 				delete func;
 				return ret;
 			}
-			static unsigned long __threadh_win_call win_target(void * imp)
+			static DWORD __threadh_win_call win_target(void * imp)
 			{
 				args * func = reinterpret_cast<args *>(imp);
 				auto ret = func->addr(func->arg);
 				delete func;
-				return reinterpret_cast<unsigned long>(ret);
+				return reinterpret_cast<DWORD>(ret);
 			}
 			static void * posix_target(void * imp)
 			{
